Add MAGICC7 tests to pymagicc
