+++ tests/test_api.py
@@ -21,32 +21,6 @@ def magicc_base():
     yield MAGICCBase()
 
 
-@pytest.fixture(scope="function", params=[MAGICC6, MAGICC7])
-def package(request):
-    MAGICC_cls = request.param
-    p = MAGICC_cls()
-
-    if p.executable is None or not exists(p.original_dir):
-        magicc_x_unavailable = "MAGICC {} is not available.".format(p.version)
-        env_text = "Pymagicc related variables in your current environment are: {}.".format(
-            ";".join(
-                [
-                    "{}: {}".format(k, v)
-                    for (k, v) in environ.items()
-                    if k.startswith("MAGICC_")
-                ]
-            )
-        )
-        env_help = "If you set MAGICC_EXECUTABLE_X=/path/to/MAGICCX/binary then you will be able to run the tests with that binary for MAGICC_X."
-        pytest.skip("\n".join([magicc_x_unavailable, env_text, env_help]))
-    p.create_copy()
-    root_dir = p.root_dir
-    yield p
-    # Perform cleanup after tests are complete
-    p.remove_temp_copy()
-    assert not exists(root_dir)
-
-
 def write_config(p):
     emis_key = "file_emissionscenario" if p.version == 6 else "FILE_EMISSCEN"
     outpath = join(p.run_dir, "MAGTUNE_SIMPLE.CFG")
